(1/2) c/p logMethod guts to shuffleAndTraxnsformArgsForPino skeleton
contents of logMethod are not changed at all, they are just copy-pasted to new location :)

shuffleAndTraxnsformArgsForPino can be used to
remap any arg list to conform to pino.LogFn
